:sheep:Populating Select With JSON Via JS
Populating select with JSON file via Javascript.

This is a file created inteded for the Stackoverflow question on how to "populate select with JSON file that coded via Javascript" at https://stackoverflow.com/questions/44793069/populate-select-with-json-file-that-coded-via-javascript?noredirect=1#comment76566945_44793069. The file is the whole codes for the following question and a collaboration of codes with https://stackoverflow.com/users/769326/dork with its plunk  plnkr.co/edit/ArfHEmkeNtIJUoQUGK6H?p=preview.